requirements.txt
vastai-sdk>=0.3.0
faster-whisper>=1.1.0
flask>=3.0.0
gunicorn>=22.0.0
requests>=2.31.0
